requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
openai-whisper==20231117
pydub==0.25.1
requests==2.32.3

